fix: lint, change path file & update crate version
