+++ Cargo.toml
@@ -2,7 +2,7 @@
 edition = "2021"
 
 name = "vsmtp"
-version = "0.7.2"
+version = "0.7.3"
 license = "GPLv3"
 
 authors = ["Team viridIT <https://viridit.com/>"]
@@ -14,16 +14,16 @@ publish = false
 
 [dependencies]
 log = "0.4.14"
-log4rs = { version = "1.0" }
+log4rs = { version = "1.0.0" }
 
-clap = "2.34.0"
+clap = { version = "3.0.4", features = ["derive"] }
 
 libc = "0.2.112"
 
-serde = { version = "1.0.132", features = ["derive"] }
-serde_derive = "1.0.132"
+serde = { version = "1.0.133", features = ["derive"] }
+serde_derive = "1.0.133"
 serde_with = { version = "1.11.0", features = ["macros"] }
-serde_json = "1.0.73"
+serde_json = "1.0.74"
 
 toml = "0.5.8"
 humantime = "2.1.0"
@@ -32,16 +32,15 @@ addr = { version = "0.15.2" } # features = ["net"] } # create some issue..
 regex = "1.5.4"
 iprange = "0.6.6"
 ipnet = "2.3.1"
-chrono = "0.4"
+chrono = "0.4.19"
 
-rhai = { version = "1.1.0", features = ["unchecked", "sync"] }
+rhai = { version = "1.3.0", features = ["unchecked", "sync"] }
 
 # the 0.10 release is more stable than the 0.9 and enable async support.
 lettre = "0.10.0-rc.4"
 
 async-trait = "0.1.52"
 lazy_static = "1.4"
-async-recursion = "0.3.2"
 
 rustls = { version = "0.20.2", features = ["tls12", "logging"] }
 rustls-pemfile = "0.2.1"
