+++ src/config/server_config.rs
@@ -1,6 +1,6 @@
 use serde_with::{serde_as, DisplayFromStr};
 
-use crate::{mailprocessing::state::StateSMTP, resolver::DataEndResolver, server::ServerVSMTP};
+use crate::{resolver::DataEndResolver, server::ServerVSMTP, smtp::state::StateSMTP};
 
 use super::custom_code::{CustomSMTPCode, SMTPCode};
