+++ src/server.rs
@@ -18,8 +18,7 @@ use crate::config::default::DEFAULT_CONFIG;
 use crate::config::log::RECEIVER;
 use crate::config::server_config::{ServerConfig, TlsSecurityLevel};
 use crate::connection::Connection;
-use crate::mailprocessing::io_service::IoService;
-use crate::mailprocessing::state::StateSMTP;
+use crate::io_service::IoService;
 use crate::resolver::DataEndResolver;
 use crate::smtp::code::SMTPReplyCode;
 use crate::transaction::Transaction;
