+++ src/smtp/mod.rs
@@ -16,3 +16,4 @@
 **/
 pub mod code;
 pub mod event;
+pub mod state;
